Add .nvmrc.
nvm is a piece of software which manages NodeJS and NPM versions.
An .nvmrc specifies the version of Node to use. This is not really
required; we are not actually using a Node webserver to host the site,
just to build it. But it's convenient to have around in case you
already are using nvm for other things.